% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{abstract_graph}
\alias{abstract_graph}
\title{Create an abstract graph object}
\usage{
abstract_graph(subclass, ...)
}
\arguments{
\item{subclass}{Desired subclass (character).}

\item{...}{Other arguments to pass to \code{list()}. See
\code{\link[=rtweet_graph]{rtweet_graph()}} for an example.}
}
\description{
Could be an actual graph object, or a graph such as the Twitter
following network defined implicitly via API requests, etc.
The abstract graph is just a list with \code{abstract_graph} class
and your desired subclass.
}
