% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{addCol}
\alias{addCol}
\title{Adds a new column to data frame}
\usage{
addCol(df, colData, colName = NULL, overwrite = FALSE)
}
\arguments{
\item{df}{A data frame}

\item{colData}{Vector. Must have same number of elements as the rows in df containing the data for the new column.}

\item{colName}{Character vector. Name for the new column. If NULL, make.names() will be used to give the column a name.}

\item{overwrite}{Boolean. Overwrite any existing data in the column named. Ignored if \code{colName} is not specified.}
}
\value{
The data frame with a new column added.
}
\description{
Adds a column to a dataframe. New column will be the last column in the data frame. Note, this should not be used to add columns from another data frame.
}

