% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-rtweet.R
\name{appr_rtweet}
\alias{appr_rtweet}
\title{Convenience function to run aPPR via rtweet}
\usage{
appr_rtweet(seeds, ...)
}
\arguments{
\item{seeds}{Screen names or user IDs of seed nodes
in the Twitter graph, as a character vector.}

\item{...}{
  Arguments passed on to \code{\link[=appr]{appr}}
  \describe{
    \item{\code{graph}}{An \code{\link[=abstract_graph]{abstract_graph()}} object, such as those created by
\code{\link[=rtweet_graph]{rtweet_graph()}} and \code{\link[=twittercache_graph]{twittercache_graph()}}, and \code{\link[igraph:igraph]{igraph::igraph()}}
object. This argument is required.}
    \item{\code{alpha}}{Teleportation constant. The teleportation constant is the
probability of returning to a seed node at each node transition.
\code{alpha} must be a valid probabilty; that is, between zero and one.
Defaults to \code{0.15}. This is the inverse of the "dampening factor"
in the original PageRank paper, so \code{alpha = 0.15} corresponds
to a dampening factor of \code{0.85}. Runtime is proportional to
\code{1 / (epsilon * alpha)}, so small \code{alpha} can result in long
runtimes.}
    \item{\code{epsilon}}{Desired accuracy of approximation. \code{epsilon} must be
a valid probabilty; that is, between zero and one. Defaults to \code{1e-6}.
Runtime is proportional to \code{1 / (epsilon * alpha)}, so small \code{epsilon}
can result in long runtimes.}
    \item{\code{tau}}{Regularization term. Additionally inflates the in-degree
of each observation by this term by performing the degree
adjustment described in Algorithm 3 and Algorithm 4, which
are described in \code{vignette("Mathematical details")}. Defaults to
\code{NULL}, in which case \code{tau} is set to the average in-degree of
the observed nodes. In general, setting it's reasonable to
set \code{tau} to the average in-degree of the graph.}
    \item{\code{verbose}}{Logical indicating whether to report on the algorithms
progress. Defaults to \code{TRUE}.}
  }}
}
\description{
Convenience function to run aPPR via rtweet
}
