% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-dev.R
\name{apprAdd}
\alias{apprAdd}
\title{Stages files for the next Git commit}
\usage{
apprAdd(..., pkgPath = getOption("appr.path"))
}
\arguments{
\item{...}{Path(s) to files to stage}

\item{pkgPath}{Path to appr's base directory. Assumed that the git repository is located there. Default value set in global options via packages \code{onLoad()}}
}
\description{
Adds specified files for the next commit
}

