% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-dev.R
\name{buildAppr}
\alias{buildAppr}
\title{Create package 'appr' from component files}
\usage{

  buildAppr(srcEnviron = "/Users/aaronwatanabe/Dropbox/Academic Documents/appr",
  srcFunctions = "functions", srcTests = "tests",
  package = "/Users/aaronwatanabe/Documents/codeDev/appr", dev = TRUE)
}
\arguments{
\item{srcEnviron}{Location of directory containing raw code.}

\item{srcFunctions}{Name of directory containing package functions}

\item{srcTests}{Name of directory containing testthat files}

\item{package}{Location of directory in which to build package}

\item{dev}{Logical. True if include developer functions in package build}
}
\value{
Returns \code{TRUE} on successfully building the package
}
\description{
Development function to build the package from the working files. Does not run tests
}

