% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{check}
\alias{check}
\title{Check if a node an abstract graph is acceptable for inclusion in PPR}
\usage{
check(graph, nodes)

check(graph, nodes)
}
\arguments{
\item{graph}{A graph object.}

\item{nodes}{The name(s) of node(s) in `graph` as a character vector.}

\item{node}{The name of a node as a character vector.}
}
\value{
Returns `TRUE` if information on `node` is available, and
  `node` has at least one outgoing edge.

The subset of `nodes` that are acceptable for inclusion. This
  can be a character vector of length zero if necessary.
}
\description{
Inclusion criteria:
}
\details{
- At least one outgoing edge
  - Can get in degree and out degree of node
  - Can get all nodes connected to `node` / the 1-hop neighbhorhood
}
