% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{dropCol}
\alias{dropCol}
\title{Drops a column}
\usage{
dropCol(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Objects which can be coerced to a vector of the column names to be dropped}
}
\value{
The data frame \code{df} without column \code{colName}
}
\description{
Removes a column from a data frame
}

