% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet-utilities.R
\name{find_token}
\alias{find_token}
\title{Find a token with remaining API requests}
\usage{
find_token(query, rate = 1, break_time = 2, strategy = "random")
}
\arguments{
\item{query}{type of API request. options in clude ...}

\item{rate}{rate limit to ask (for use)}

\item{break_time}{time interval to wait, in minutes}

\item{strategy}{'random' for uniformly at random,}
}
\value{
TODO
}
\description{
Find a token with remaining API requests
}
