% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-misc.R
\name{isInteger}
\alias{isInteger}
\title{Checks whether the given object is an integer number}
\usage{
isInteger(x)
}
\arguments{
\item{x}{A vector to test for integerness}
}
\value{
A logical vector the same length as \code{x} in which the \code{i}th element is TRUE if the \code{i}th element of \code{x} is an integer.
}
\description{
Checks whether the given object is an integer number. This is differnet behavior from \code{\link[base]{is.integer}}, which checks whether the variable type is integer. This often will not be useful.
}

