% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-io.R
\name{loadCSV}
\alias{loadCSV}
\title{Load CSV file}
\usage{
loadCSV(filePath, headers = TRUE, factorLabels = FALSE,
  encodingFile = "UTF-8", encoding = "UTF-8", ...)
}
\arguments{
\item{filePath}{Path to CSV file to load into memory.}

\item{headers}{Is the first row of the file column names?}

\item{factorLabels}{Convert variables with values as a factors with those levels.}

\item{encodingFile}{Encoding of the original file (UTF-8, macintosh, etc)}

\item{encoding}{Marks imported strings as known to have a particular coding. Does not change the encoding itself, which will be automatically set to the encoding R is using on the local system. See 'Encoding.'}

\item{...}{Additional arguments to pass to \code{\link[utils]{read.csv}}.}
}
\value{
A data frame with the contents of the CSV file.
}
\description{
Wrapper function around utils::read.csv() to load CSV (comma separated values) files into memory as data frames.
}
\section{Encoding}{

Encodings refer to the way that characters are stored in memory (or more specifically how R translates the zeros and ones in memory into characters). Unfortunately, there are many encoding systems. R will use one encoding, which \href{https://stat.ethz.ch/pipermail/r-sig-mac/2007-March/003733.html}{at least on macs is \code{UTF-8}}, but you may be reading data files from another. R will automatically convert data to its encoding system on import, but may need to know what those are. Use \code{encodingFile} to specify the encoding of the file you're importing from. Generally, you should not need to change \code{encoding}, unless R is using \code{Laitn-1} (perhaps the standard for Windows) or a different encoding on your machine.
}

