% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{mapCols}
\alias{mapCols}
\title{Maps function onto each element of the specified columns}
\usage{
mapCols(df, fun, cols = c(), ...)
}
\arguments{
\item{df}{A dataframe.}

\item{fun}{A function to apply to each element of the selected columns. Elements of these columns will be the first argument passed to this function.}

\item{cols}{A vector containing the names (as strings) or index numbers of the columns with elements to which to apply \code{fun}.}

\item{...}{Additional arguments to pass to fun}
}
\value{
A dataframe of the same dimensions as \code{df} with \code{cols} containing the results of applying \code{fun} to each element.
}
\description{
Maps \code{fun} to each element of the the columns of \code{df} specified in \code{...}. Returns the original data frame with the values of specified columns updated with the values produced from mapping the function
}
\section{Details}{

\code{mapCols} works by folding over the set of columns to manipulate. For each column in \code{cols}, runs \code{fun} over its elements using sapply and replacing the column's original contents with the output.
}

