% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcardio.r
\name{model_cardio}
\alias{model_cardio}
\title{Title}
\usage{
model_cardio(age, sex, thalach)
}
\arguments{
\item{age}{Edad del paciente}

\item{sex}{Genero del paciente ("Female" or "Male)}

\item{thalach}{Presion sanguinea del paciente}
}
\value{

}
\description{
Title
}
\examples{
model_cardio(30,"Female",100)
}
