% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{neighborhood}
\alias{neighborhood}
\title{Get the neighborhood of a node in a graph}
\usage{
neighborhood(graph, node)
}
\arguments{
\item{graph}{A graph object.}

\item{nodes}{The name of a single node in `graph` as a character vector.}
}
\value{
A character vector of all nodes in `graph` connected such that
  there is an outgoing edge for `node` to those nodes. This should
  never be empty, as `neighborhood()` should not be called on nodes
  that fail `check()`.
}
\description{
That is, find all nodes connected to `node` by an outgoing edge.
This function is memorized to avoid making repeated API queries.
}
