% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-neocache.R
\name{neocache_graph}
\alias{neocache_graph}
\title{Create an abstract representation of the Twitter friendship graph}
\usage{
neocache_graph(cache_name = "aPPR", retryonratelimit = TRUE)
}
\arguments{
\item{attempts}{The number of times to attempt an API request before
moving on or erroring.}
}
\description{
Signifies that \code{aPPR} should query the Twitter friendship graph via
\code{neocache}.
}
\examples{

\dontrun{


test_ids <- c("780429268866052096", "1191642560")

graph <- neocache_graph()

check(graph, test_ids)
node_degrees(graph, test_ids)
neighborhood(graph, test_ids[1])


}

}
