% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{node_degrees}
\alias{node_degrees}
\title{Get the in-degree and out-degree of nodes in an abstract graph}
\usage{
node_degrees(graph, nodes)
}
\arguments{
\item{graph}{A graph object.}

\item{nodes}{The name(s) of node(s) in \code{graph} as a character vector.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with one row for every node in \code{nodes} and
two columns: \code{in_degree} and \code{out_degree}.frame with one
}
\description{
This function is only called nodes that have been \code{\link[=check]{check()}}'d. It is
safe to assume that \code{nodes} is non-empty. (TODO: check this!)
}
