% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{recodeCol}
\alias{recodeCol}
\title{Recodes elements of a column according to a conversion table}
\usage{
recodeCol(df, colName, convertTable, destName = NULL)
}
\arguments{
\item{df}{A data frame}

\item{colName}{Name of column to code}

\item{convertTable}{Data frame containing conversation scores. Should contain two columns: oldValue and newValue. The table must have as many rows as there are unique values in \code{colName}. Each row should correspond an oldValue to the newly coded value.}

\item{destName}{Character vector. Name of new column to write recoded data to. May not already ben the name of a column in \code{df}. If NULL, recoded data will be written over the original contents in \code{colName}.}
}
\value{
The original data frame with the new column \code{scoreName} containing the results of scoring \code{colName} with \code{scoreTable}
}
\description{
Recodes elements of a column according to a conversion table and writes them to a new column. Will resort the rows in \code{df}.
}

