% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{renameCol}
\alias{renameCol}
\title{Renames column in data frame}
\usage{
renameCol(df, oldName = "", newName = "")
}
\arguments{
\item{df}{A dataframe.}

\item{oldName}{The name of the column to be renamed.}

\item{newName}{The new name for the column.}
}
\value{
The dataframe with the column renamed.
}
\description{
Wrapper function to make column renaming easier.
}

