% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{replaceValues}
\alias{replaceValues}
\title{Replace values in columns in specified rows}
\usage{
replaceValues(df, searchIndex, ...)
}
\arguments{
\item{df}{A dataframe}

\item{searchIndex}{A logical vector with length equal to the number of rows in \code{df} in which the \code{i}th element is 'TRUE' if the \code{i}th row in \code{df} should be subject to replacement. Can be generated using \code{\link{searchDF}}.}

\item{...}{Replacement columns and values, passed as consequetive arguments with the column name as the first argument and the new value as the second.}
}
\value{
A new copy up \code{df} with the specified rows and columns updated.
}
\description{
Updates named column(s) in \code{df} with given value(s) for the rows indicated by \code{searchIndex} and returns new copy of \code{df}.
}

