% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{searchDF}
\alias{searchDF}
\title{Searchs given data frame for rows in which the given column has the specified value}
\usage{
searchDF(df, col, colValue)
}
\arguments{
\item{df}{A data frame to search}

\item{col}{Name or index of column to search for \code{colValue}.}

\item{colValue}{Value in \code{column} to search for. Must be the type of \code{col}.}
}
\description{
Searchs dataframe \code{df} for rows containing \code{colValue} in \code{col}. Returns an index vector (i.g. logical vector that is the legnth of the number of rows of \code{df}) where the \code{i}th element is TRUE if the \code{i}th row of \code{df} meets the search criterion
}

