% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.Tracker}
\alias{update.Tracker}
\title{Update a Tracker object}
\usage{
\method{update}{Tracker}(object, ..., epsilon, verbose = TRUE)
}
\arguments{
\item{object}{The \code{Tracker} object to update.}

\item{...}{Ignored. Passing arguments to \code{...} results in a warning.}

\item{epsilon}{Desired accuracy of approximation. \code{epsilon} must be
a valid probabilty; that is, between zero and one. Defaults to \code{1e-6}.
Runtime is proportional to \code{1 / (epsilon * alpha)}, so small \code{epsilon}
can result in long runtimes.}

\item{verbose}{Logical indicating whether to report on the algorithms
progress. Defaults to \code{TRUE}.}
}
\value{

}
\description{
Typically because results are insufficiently precise.
}
\details{
At the moment, only supports changing \code{epsilon}. If there is interest,
we can consider allowing updates to \code{tau}, \code{alpha} and \code{seeds} in the
future.
}
