% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watanabe-df-manipulation.R
\name{updateDF}
\alias{updateDF}
\title{Carries out multiple replacements of values in a dataframe}
\usage{
updateDF(df, updates)
}
\arguments{
\item{df}{The data frame to update.}

\item{updates}{A dataframe containing information that identifies columns and their replacement values as well as which rows should be updated. Format is specified below.}
}
\description{
Carries out multiple replacements of values in a dataframe
}
\section{Update Table}{

The update table contains information about which rows to update and the columns in \code{df} and values they should be updated with.
}

