% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{abstract_graph}
\alias{abstract_graph}
\title{Create an abstract graph object}
\usage{
abstract_graph(subclass, ...)
}
\arguments{
\item{subclass}{TODO}

\item{...}{TODO}
}
\value{
Returns `TRUE` if information on `node` is available, and
  `node` has at least one outgoing edge.
}
\description{
Could be an actual graph object, or a graph such as the Twitter
following network defined implicitly via API requests, etc.
}
