% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPPR-package.R
\docType{package}
\name{aPPR-package}
\alias{aPPR}
\alias{aPPR-package}
\title{aPPR: Approximate Personalized Page Rank}
\description{
aPPR helps you calculate approximate
    personalized pageranks from large graphs, including those that can
    only be queried via an API.  aPPR additionally performs degree
    correction and regularization, allowing users to recover blocks from
    stochastic blockmodels.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RoheLab/aPPR}
  \item Report bugs at \url{https://github.com/RoheLab/aPPR/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID})

Authors:
\itemize{
  \item Fan Chen
  \item Karl Rohe
}

}
\keyword{internal}
