% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPPR-package.R
\docType{package}
\name{aPPR-package}
\alias{aPPR}
\alias{aPPR-package}
\title{aPPR: Approximate Personalized PageRank}
\description{
Calculates approximate and regularized personalized PageRank vectors for massive graphs, including those that can only be queried via an API. Regularization allows discovery of community structure under some stochastic block models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RoheLab/aPPR}
  \item Report bugs at \url{https://github.com/RoheLab/aPPR/issues}
}

}
\author{
\strong{Maintainer}: Alex Hayes \email{alexpghayes@gmail.com} (\href{https://orcid.org/0000-0002-4985-5160}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Fan Chen \email{fan.chen@wisc.edu} (\href{https://orcid.org/0000-0003-4508-6023}{ORCID})
  \item Karl Rohe
}

}
\keyword{internal}
