% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appr-package.R
\docType{package}
\name{appr}
\alias{appr}
\alias{appr-package}
\title{appr - Application framework for R}
\description{
A framework for building pluggable command line R applications based on
descriptor files. Automatically generates a command line interface that
supports commands from installed packages by configuring them in a
descriptor file. Supports options at the application and command level,
including configuration files.
}
