% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPPR.R
\name{appr}
\alias{appr}
\title{Approximate personalized pageranks}
\usage{
appr(graph, seeds, alpha = 0.15, epsilon = 1e-06, adjust = TRUE,
  tau = 0, ...)
}
\arguments{
\item{graph}{TODO}

\item{seeds}{TODO}

\item{alpha}{TODO}

\item{epsilon}{TODO}

\item{adjust}{TODO}

\item{tau}{TODO}

\item{...}{Ignored. Passing arguments to `...` results in a warning.}
}
\value{
TODO
}
\description{
Approximate personalized pageranks
}
\examples{

#### on a concrete, local igraph object

library(igraph)

set.seed(27)

ig <- sample_pa(100)

gcon <- igraph_connection(ig)

in_degree(gcon, "1")
out_degree(gcon, "1")
neighborhood(gcon, "1")

##### on the twitter graph via rtweet

# TODO

##### on the twitter graph via twittercache

# TODO

}
