% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPPR.R, R/graph-igraph.R, R/graph-rtweet.R
\name{appr}
\alias{appr}
\alias{appr.igraph}
\alias{appr.rtweet_graph}
\title{Approximate personalized pageranks}
\usage{
appr(graph, seeds, alpha = 0.15, epsilon = 1e-06, tau = NULL, ...)

\method{appr}{igraph}(graph, seeds, alpha = 0.15, epsilon = 1e-06, tau = NULL, ...)

\method{appr}{rtweet_graph}(graph, seeds, alpha = 0.15, epsilon = 1e-06, tau = NULL, ...)
}
\arguments{
\item{graph}{An [abstract_graph()] object, such as those created by
[rtweet_graph()] and [twittercache_graph()], and [igraph::igraph()]
object. This argument is required.}

\item{seeds}{A character vector of seeds for the personalized pagerank.
The personalized pagerank will return to each of these seeds with
probability `alpha` at each node transition. At the moment,
all seeds are given equal weighting. This argument is required.}

\item{alpha}{Teleporting factor. The teleportation factor is the
probability of returning to a seed node at each node transition.
Defaults to `0.15`. This is the inverse of the "dampening factor"
in the original PageRank paper, so `alpha = 0.15` corresponds
to a dampening factor of `0.85`.}

\item{epsilon}{Desired accuracy of approximation. Small `epsilon`
can result in long runtimes. Defaults to `1e-6`.}

\item{tau}{Regularization term. Additionally inflates the in degree
of each observation by this term by performing the degree
adjustment described in Algorithm 3 and Algorithm 4, which
are described in `vignette("Mathematical details")`. Defaults to
`NULL`, in which case `tau` is set to the average in degree of
the observed nodes. In general, setting it's reasonable to
set `tau` to the average degree of the graph.}

\item{...}{Ignored. Passing arguments to `...` results in a warning.}
}
\value{
A [tibble::tibble()] with the following columns:

  - `name`: Name of a node (character).
  - `p`: Estimated personalized pagerank of a node.
  - `r`: Estimated error of pagerank estimate for a node.
  - `in_degree`: Number of incoming edges to a node.
  - `out_degree`: Number of outcoming edges from a node.
  - `degree_adjusted`: The personalized pagerank divided by the
    node in-degree.
  - `regularized`: The personalized pagerank divide by the node
    in-degree plus `tau`.

When `graph` is an

  - `rtweet_graph` get `user_id` not screen names
}
\description{
Approximate personalized pageranks
}
\details{
Note that, due to the inspection paradox, the average observed
degree of the network will be higher than the average degree of
the entire network. This means that the default value for `tau`
many be a bit high. In practice, we do not observe that the
small changes in the value of `tau` substantively changes results,
you really just need to regularize enough so that low in degree
nodes don't look overly important after degree correction.
}
\examples{

library(aPPR)
library(igraph)

set.seed(27)

graph <- rtweet_graph()

\dontrun{
appr(graph, "alexpghayes")
}

graph2 <- sample_pa(100)

appr(graph2, seeds = "5")


library(igraph)

set.seed(27)

graph <- sample_pa(100)

appr(graph, seeds = "5")

}
