% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-rtweet.R
\name{appr_rtweet}
\alias{appr_rtweet}
\title{Convenience function to run aPPR via rtweet}
\usage{
appr_rtweet(seeds, ...)
}
\arguments{
\item{seeds}{TODO}

\item{...}{
  Arguments passed on to \code{\link[=appr.rtweet_graph]{appr.rtweet_graph}}
  \describe{
    \item{\code{graph}}{An [abstract_graph()] object, such as those created by
[rtweet_graph()] and [twittercache_graph()], and [igraph::igraph()]
object. This argument is required.}
  }}
}
\description{
Convenience function to run aPPR via rtweet
}
