% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{check}
\alias{check}
\title{Check if a node an abstract graph is acceptable for inclusion in PPR}
\usage{
check(graph, node)
}
\arguments{
\item{graph}{A graph object.}

\item{node}{The name of a node as a character vector.}
}
\value{
Returns `TRUE` if information on `node` is available, and
  `node` has at least one outgoing edge.
}
\description{
Check if a node an abstract graph is acceptable for inclusion in PPR
}
