% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{check}
\alias{check}
\title{Check if a node an abstract graph is acceptable for inclusion in PPR}
\usage{
check(graph, nodes)
}
\arguments{
\item{graph}{A graph object.}

\item{nodes}{The name(s) of node(s) in \code{graph} as a character vector.}
}
\value{
The subset of \code{nodes} that are acceptable for inclusion. This
can be a character vector of length zero if necessary.
}
\description{
Inclusion criteria:
}
\details{
\itemize{
\item At least one outgoing edge
\item Can get in degree and out degree of node
\item Can get all nodes connected to \code{node} / the 1-hop neighbhorhood
}
}
