% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-graph.R
\name{neighborhood}
\alias{neighborhood}
\title{Get the neighborhood of a node in a graph}
\usage{
neighborhood(graph, node)
}
\arguments{
\item{graph}{A graph object.}

\item{node}{The name of a single node in \code{graph} as a character vector.}
}
\value{
A character vector of all nodes in \code{graph} connected such that
there is an outgoing edge for \code{node} to those nodes. This should
never be empty, as \code{neighborhood()} should not be called on nodes
that fail \code{check()}.
}
\description{
That is, find all nodes connected to \code{node} by an outgoing edge.
This function is memorized to avoid making repeated API queries.
}
