% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-neocache.R
\name{neocache_graph}
\alias{neocache_graph}
\title{Create an abstract representation of the Twitter friendship graph}
\usage{
neocache_graph(cache_name = "aPPR", retryonratelimit = TRUE)
}
\arguments{
\item{cache_name}{The name of the \code{neocache} to use to save data into.}

\item{retryonratelimit}{If \code{TRUE}, and a rate limit is exhausted, will wait
until it refreshes. Most Twitter rate limits refresh every 15 minutes.
If \code{FALSE}, and the rate limit is exceeded, the function will terminate
early with a warning; you'll still get back all results received up to
that point. The default value, \code{NULL}, consults the option
\code{rtweet.retryonratelimit} so that you can globally set it to \code{TRUE},
if desired.

If you expect a query to take hours or days to perform, you should not
rely soley on \code{retryonratelimit} because it does not handle other common
failure modes like temporarily losing your internet connection.}
}
\description{
Signifies that \code{aPPR} should query the Twitter friendship graph via
\code{neocache}.
}
\examples{

\dontrun{


test_ids <- c("780429268866052096", "1191642560")

graph <- neocache_graph()

check(graph, test_ids)
node_degrees(graph, test_ids)
neighborhood(graph, test_ids[1])


}

}
