% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-rtweet.R
\name{rtweet_graph}
\alias{rtweet_graph}
\title{Create an abstract representation of the Twitter friendship graph}
\usage{
rtweet_graph(attempts = 5)
}
\arguments{
\item{attempts}{The number of times to attempt an API before
moving on or erroring.}
}
\description{
Signifies that we should query the Twitter friendship graph via
`rtweet`.
}
