% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-rtweet.R
\name{rtweet_graph}
\alias{rtweet_graph}
\title{Create an abstract representation of the Twitter friendship graph}
\usage{
rtweet_graph(attempts = 5)
}
\arguments{
\item{attempts}{The number of times to attempt an API request before
moving on or erroring.}
}
\description{
Signifies that we should query the Twitter friendship graph via
\code{rtweet}. In practice, we actually use \code{socialsampler}, so you
can use multiple tokens at once. See \code{\link[socialsampler:register_token]{socialsampler::register_token()}}
for details.
}
