% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.Tracker}
\alias{update.Tracker}
\title{Update a Tracker object}
\usage{
\method{update}{Tracker}(object, ..., epsilon, verbose = FALSE)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Typically because results are insufficiently precise.
}
\details{
At the moment, only supports changing \code{epsilon}. If there is interest,
we can consider allowing updates to \code{tau}, \code{alpha} and \code{seeds} in the
future.
}
